+++ server/components/dashboard/dashboardController.ts
@@ -6,9 +6,18 @@ import { ReturnSuccess } from '../../services/httpHandling/httpHandlingService';
 import { badImplementation } from '@hapi/boom';
 
 export class DashboardController {
-  public static async getMetrics(_req: FastifyRequest, res: FastifyReply<any>): Promise<any> {
+  public static async getCheckers(_req: FastifyRequest, res: FastifyReply<any>): Promise<any> {
     try {
-      const result = await DashboardService.getMetrics();
+      const result = await DashboardService.getCheckers();
+      ReturnSuccess(res, undefined, { data: result });
+    } catch (error) {
+      HandlingBoom(badImplementation(), res);
+    }
+  }
+
+  public static async getMetricsByCheckerId(req: FastifyRequest, res: FastifyReply<any>): Promise<any> {
+    try {
+      const result = await DashboardService.getMetricsByCheckerIdCache(req.params.id);
       ReturnSuccess(res, undefined, { data: result });
     } catch (error) {
       HandlingBoom(badImplementation(), res);