+++ server/components/dashboard/dashboardRoutes.test.ts
@@ -54,7 +54,7 @@ describe('DashboardRoutes', () => {
     fastify.close().then(() => done());
   });
 
-  describe('GET /dashboard/metrics', () => {
+  describe('GET /dashboard/checkers', () => {
     beforeAll(async done => {
       const data = {
         ms: 2,
@@ -77,17 +77,57 @@ describe('DashboardRoutes', () => {
       done();
     });
 
-    it('should return all data for public user', async () => {
+    afterAll(async done => {
+      await fastify.inject({
+        method: 'PUT',
+        url: `/checkers/${checker._id}/stop`,
+        ...authHeaders(token),
+      });
+      done();
+    });
+
+    it('should return all checkers with groups', async () => {
       const res = await fastify.inject({
         method: 'GET',
-        url: '/dashboard/metrics',
+        url: '/dashboard/checkers',
+      });
+      expect(res.statusCode).toEqual(200);
+      expect(JSON.parse(res.body)).toMatchObject({ message: 'Success' });
+      expect(Object.keys(JSON.parse(res.body).data)).toMatchObject(['groups', 'checkers']);
+    });
+  });
+
+  describe('GET /dashboard/metrics/:id', () => {
+    beforeAll(async done => {
+      const data = {
+        ms: 2,
+        statusCode: 200,
+        checkerId: checker._id,
+      };
+
+      await fastify.inject({
+        method: 'PUT',
+        url: `/checkers/${checker._id}/start`,
+        ...authHeaders(token),
+      });
+
+      await fastify.inject({
+        method: 'POST',
+        url: '/metrics',
+        ...authHeaders(token),
+        body: data,
+      });
+      done();
+    });
+
+    it('should return all data for dedicated checkers', async () => {
+      const res = await fastify.inject({
+        method: 'GET',
+        url: `/dashboard/metrics/${checker._id}`,
       });
       expect(res.statusCode).toEqual(200);
       expect(JSON.parse(res.body)).toMatchObject({ message: 'Success' });
       expect(Object.keys(JSON.parse(res.body).data)).toMatchObject([
-        'groups',
-        'checkers',
-        'metricsStatusCode',
         'metricsMs',
         'metricsStatusCodeSum',
         'metricsStatusCodeSumKeys',